1win_ar/requirements.txt
fastapi==0.123.5
uvicorn==0.38.0
playwright==1.56.0
structlog==24.1.0
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-telegram-bot>=21.0